refactor: 게이트웨이 서비스 경로 수정 및 컨벤션 적용
