+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,4 +1,5 @@
 ## ✅ PR 유형
+
 어떤 변경 사항이 있었나요?
 
 - [ ] 새로운 기능 추가
@@ -12,6 +13,7 @@
 ---
 
 ## ✏️ 작업 내용
+
 > [!CAUTION]
 > 스크린샷 첨부는 **필수**입니다!
 > **첨부 후 해당 콜아웃은 지워주세요!!**
@@ -19,6 +21,7 @@
 ---
 
 ## 🔗 관련 이슈
+
 - close #
 
 ---
