+++ README.md
@@ -1,2 +1,3 @@
 # gateway-service
+
 Trendist 백엔드 gateway-service 레포지토리입니다.
