+++ src/main/java/com/trendist/gateway_service/GatewayServiceApplication.java
@@ -10,9 +10,13 @@ import org.springframework.web.reactive.function.client.WebClient;
 @SpringBootApplication
 public class GatewayServiceApplication {
 
+	public static void main(String[] args) {
+		SpringApplication.run(GatewayServiceApplication.class, args);
+	}
+
 	@Bean
 	@LoadBalanced
-	public WebClient.Builder loadBalancedWebClientBuilder(){
+	public WebClient.Builder loadBalancedWebClientBuilder() {
 		final WebClient.Builder builder = WebClient.builder();
 		return builder;
 	}
@@ -21,9 +25,5 @@ public class GatewayServiceApplication {
 	@Bean
 	public RestClient.Builder restClientBuilder() {
 		return RestClient.builder();
-	}
-
-	public static void main(String[] args) {
-		SpringApplication.run(GatewayServiceApplication.class, args);
 	}
 }
