--- src/main/java/com/trendist/gateway_service/exception/AccessTokenExpiredException.java
@@ -1,8 +1,0 @@
-// AccessTokenExpiredException.java
-package com.trendist.gateway_service.exception;
-
-public class AccessTokenExpiredException extends RuntimeException {
-    public AccessTokenExpiredException() {
-        super("Access token has expired");
-    }
-}
