--- src/main/java/com/trendist/gateway_service/exception/NotExistsAuthorization.java
@@ -1,7 +1,0 @@
-package com.trendist.gateway_service.exception;
-
-public class NotExistsAuthorization extends RuntimeException {
-    public NotExistsAuthorization() {
-        super("Authorization header does not exist");
-    }
-}
