+++ src/main/java/com/trendist/gateway_service/global/exception/AccessTokenExpiredException.java
@@ -1,0 +1,7 @@
+package com.trendist.gateway_service.global.exception;
+
+public class AccessTokenExpiredException extends RuntimeException {
+	public AccessTokenExpiredException() {
+		super("Access token has expired");
+	}
+}
