+++ src/main/java/com/trendist/gateway_service/global/exception/NotExistsAuthorization.java
@@ -1,0 +1,7 @@
+package com.trendist.gateway_service.global.exception;
+
+public class NotExistsAuthorization extends RuntimeException {
+	public NotExistsAuthorization() {
+		super("Authorization header does not exist");
+	}
+}
